ai/requirements.txt
ultralytics>=8.2.0
albumentations>=1.4.2
opencv-python-headless>=4.9.0.80
torch>=2.1
tqdm>=4.66
PyYAML>=6.0.1
onnxruntime>=1.17.0
onnxconverter-common>=1.13.0
Pillow>=9.5.0